Update patch set 1
Patch Set 1:

> I think this would break the release scripts
What are these?

> I don't like adding fake dependencies to support that. 
Actually, we could argue that //third_party is a dependency of the project! As are all other directories/modules.

> I think the dependency tree should reflect the real dependency relations on what's checked out.
Why adding "on what's checked out"? :)

> What I would prefer is parsing the local DEPS file and "merging" it into the current dependency.
Why not, but it would require a substantial amount of refactoring/tweaking, it seems. E.g. extracting Dependency().ParseDepsFile() as a standalone function.
Is it worth it?

Patch-set: 1
Reviewer: Gerrit User 1286313 <1286313@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>