Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > I think this would break the release scripts
> What are these?
I don't know the details, but they're some recipes that create the release branches.
I know because I've broken them before doing things like these. ;)
I think you should talk with mmoss@ to check how a change like this would affect it.

> > I don't like adding fake dependencies to support that. 
> Actually, we could argue that //third_party is a dependency of the project! As are all other directories/modules.
Well, yes, but we could argue it is part of the current repo, so it's dependencies are the current repo dependencies.

> > I think the dependency tree should reflect the real dependency relations on what's checked out.
> Why adding "on what's checked out"? :)
> > What I would prefer is parsing the local DEPS file and "merging" it into the current dependency.
> Why not, but it would require a substantial amount of refactoring/tweaking, it seems. E.g. extracting Dependency().ParseDepsFile() as a standalone function.
> Is it worth it?
I'd like to think it is worth it.
There are a number of teams that would benefit from being able to parse a DEPS file without having it be part of a repository.
None URLs are a workaround to make things work, but they tend to make the code harder to reason about and to refactor.
Sorry if I'm pushing back too much, but I'd like to make things easier for the person that has to deal with this codebase in the future.

That said, if we do it this way, please talk with mmoss@ first, and add tests for this.

Patch-set: 1
Reviewer: Gerrit User 1167909 <1167909@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>