Update patch set 6
Patch Set 6:

I think this wouldn't be needed when [1] lands. But this seems more straightforward and less risky, so it might be worth it to land it, specially if the review of [1] would take more time.

[1] https://chromium-review.googlesource.com/c/chromium/tools/depot_tools/+/1119235

Patch-set: 6
Reviewer: Gerrit User 1167909 <1167909@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0