Update patch set 2
Patch Set 2:

Thanks for the reviews. I tested the patch in Windows and it turns out that ctypes.windll.kernel32 isn't working as intended when running in glient (vpython? some lib?) so I rewrote that part so that it works in Windows as well. 

When it fails it does so silently by not truncating progress to the width of the terminal so it's not obvious.

And yes, the python3 function would be nice. Maybe python3 is a serious option soon?

Patch-set: 2
Reviewer: Gerrit User 1114374 <1114374@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>