Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > As stated, the lint error to Catapult is not the only instance. This change was also blocking the Chromium PRESUBMIT
> 
> that wasn't stated anywhere.  you've mentioned the Chromium PRESUBMIT failure (and linked to a log), but that's the only thing you've done.  you haven't said what the Catapult error was, or linked to any example logs.  so i'm left trying to guess at what the actual problem is on the Catapult side.

Sorry for wasn't being cleared. The Catapult error is on the the first comment of the linked bug. The Chromium PRESUBMIT failure is described in the log is also similar. In both instances, it's about the missing import module and not about Catapult or Chromium code being changed.


> 
> > Is there any announcement about this pylint change
> 
> yes, a PSA already went to chromium-dev@ before the previous one landed

Patch-set: 1
Reviewer: Gerrit User 1128435 <1128435@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>