Update patch set 1
Patch Set 1: Commit-Queue+1

> Sorry for wasn't being cleared. The Catapult error is on the the first comment of the linked bug.

the linked bug is crbug.com/863669.  looking at your revert in CL:1145160, that was also the bug linked at the bottom using the Bug: style tag.  but it does look like you had an old-style BUG= line to crbug.com/865897 which has the full error message and i had missed that.  it's the same missing lazy_object_proxy issue.

> The Chromium PRESUBMIT failure is described in the log is also similar

the linked log for that was helpful, but when it was mentioned, it was as "the Chromium presubmit is failing" not "the catapult is failing the same was as this chromium presubmit failure".

at any rate, it looks like this CL should fix both since it was the same underlying problem -- the bots didn't have some new python modules installed.

Patch-set: 1
Reviewer: Gerrit User 1000768 <1000768@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1