Update patch set 3
Patch Set 3:

to be clear, your revert due to missing module imports in pylint itself was the right thing to do, so thanks for that as i wasn't around to handle it.

the point of contention was trying to find out the reason for the revert (and i think we finished up that rabbit hole), because in the past we've seen people revert linters due to new lint warnings showing up in their code (instead of fixing the new lint output, they go back to the old linter).

i've tried to make these lint upgrades not super noisy for everyone, but it's at the cost of disabling a lot of linter functionality (as you can see in the long list in pylintrc).

Patch-set: 3