Update patch set 2
Patch Set 2:

Thanks for the additional info.

I guess that means that we either (a) change the release scripts to not assume that every DEPS file that is recursed into is in a different repo, or (b) implement `include` or something like that for files in the same repo.

I still kinda feel like (a) is the right thing to do, but I feel less strongly about it than I did before.

Patch-set: 2