Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Thanks for the additional info.
> > 
> > I guess that means that we either (a) change the release scripts to not assume that every DEPS file that is recursed into is in a different repo, or (b) implement `include` or something like that for files in the same repo.
> > 
> > I still kinda feel like (a) is the right thing to do, but I feel less strongly about it than I did before.
> 
> So, to be clear: ehmaldonaldo/agable, if either/both of you still prefer (b), I will defer to you and stop arguing :).

I would pretty strongly prefer an 'include' mechanism: a DEPS file in the same repo is not a dependency, it is a source of other dependencies. I think that overloading the concept of dependency, or even just the concept of "recursedeps" (which I'd also like to get rid of and replace with a boolean inside the value of each deps entry), to include things that... aren't dependencies... is a move in the opposite direction for what I'd like in gclient.

'include' mechanisms are well-traversed territory for similar files like BUILD and OWNERS. I'd much rather do that than add additional confusion to the deps={} mechanism.

Patch-set: 2
Reviewer: Gerrit User 1002100 <1002100@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>