Update patch set 2
Patch Set 2:

(5 comments)

In general, we must make sure that we can generate a single DEPS file with all vars, deps, hooks, etc. So we can't have local vars, or deps with different versions in different included DEPS.
And I think we should only need to support the minimum we need for //src/build and nothing else.

Patch-set: 2