Update patch set 3
Patch Set 3:

Me and Yves discussed this today. I had originally imagined that most third_party/ deps used by //build would be used by //build only, but it turns out that roughly 80% of the deps used by //build are also used by //tools, //v8, //nacl and others. Yves wrote a heuristic that checks this.

Then the fundamental problem to solve is to deal with those; you have talked about that as the gclient setdep problem earlier I believe. We could keep the 80% in src/DEPS, but then we haven't gained anything, and it's confusing for chromium devs whether to put a new dep in src/DEPS or src/build/DEPS. We could also say that all deps used by //build should go in src/build/DEPS even if they're used elsewhere in Chromium. That works, but it doesn't make a lot of sense from a Chromium perspective; the lowest common denominator in that case is src/DEPS, which is where they are located today.

I don't see an obvious way to move forward, so maybe we should scrap this effort? It's possible we can copy the generated android deps list in our roll script and at least solve the most obvious problem (Android DEPS changing). Maybe we can encourage a future where as much as possible of Chromium's DEPS is generated and is safe for us to copy?

Patch-set: 3
Reviewer: Gerrit User 1002301 <1002301@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>