Update patch set 1
Patch Set 1: Code-Review+1

This weirdness doesn't really make sense to me... 
If it doesn't make sense to you either, can you add a TODO to look into this and see if any project relies on this, and if not, come up with something easier to understand?

Regardless, I think it's a good idea to test this, so we don't inadvertently change the behavior.
Thanks for writing this!

Patch-set: 1
Reviewer: Gerrit User 1167909 <1167909@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1