[gclient] Rename gclient_config param to cfg
gclient config parameters in gclient module are called "gclient_config".
This is a little verbose and "gclient_" prefix does not add much value in the
context of gclient module. Rename to just "cfg".

Also it is slightly simpler because now there is only one local variable
as opposed to gclient_config and cfg.

Bug: 694348
Change-Id: Id03970399b62489425da54da92279fe6f0194e41