Update patch set 1
Patch Set 1:

(2 comments)

Would it be feasible/a good idea to require that the accounts used for reviewers and CC are valid?
If they're not, the upload would fail and the user would be asked to file a ticket.

Patch-set: 1
Reviewer: Gerrit User 1167909 <1167909@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>