Add diff support to python git cl format and enable by default
If a .style.yapf file is not found in any parent directory the
.style.yapf file in depot tools will be used instead.

It seems that yapf will still make changes to the indentation of
the entire file even if line ranges are specified.
This is intended see https://github.com/google/yapf/issues/499
This could potentially cause some issues if there are a lot of
indentation style violations in a file or if format is accidentally run
on a pep8 formatted file in third_party without a .style.yapf file.

Yapf formatting can be disabled by using the following comments
# yapf: disable
FOO = {
    # ... some very large, complex data literal.
}

BAR = [
    # ... another large data literal.
]
# yapf: enable
See https://github.com/google/yapf


Bug:846432
Change-Id: Ib85797f4a8e1021870901ff465ec10f7e70deb87