Update patch set 5
Patch Set 5:

This change moves cipd downloads to after the hooks have been run, which breaks the Dart checkout.  Our stable SDK is downloaded with cipd, and scripts in our hooks either reference it or use the Dart executable in our checked-out SDK to execute scripts.

It seems like the problem this change was meant to fix is much smaller than the problem that not all our dependencies are checked out before the hooks run.  Remember that there can now be cipd dependencies, that are fetched with cipd, as well as git dependencies.

I can file an issue on this, if we want to discuss this further.

Ours cipd depencency looks like this, and is in the "deps" dictionary, not "hooks":

 Var("dart_root") + "/tools/sdks": {
      "packages": [
          {
              "package": "dart/dart-sdk/${{platform}}",
              "version": "version:2.1.0-dev.4.0",
          },
      ],
      "dep_type": "cipd",
  },

Patch-set: 5
CC: Gerrit User 1183500 <1183500@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1238297 <1238297@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1116069 <1116069@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>