Update patch set 5
Patch Set 5:

The problem that William describes breaks fresh checkouts of older commit of Dart. We've added a workaround but it was an incompatible change to us. Do you have any thoughts about whether this incompatible change was a feature or a bug? It would be useful if we could rely on CIPD packages being around when the hooks are run.

Patch-set: 5