metrics: Add a mechanism to notify users when we want to collect additional metrics.
When we change the version number in metrics_utils:

If the user is not a Googler, or has opted out explicitly, nothing happens, and we still don't collect metrics.

If we're collecting metrics from the user, we stop collecting metrics and display a notice telling them what has changed.
That notice will be displayed ten times, after which we will resume collecting metrics. A notice telling them we're collecting metrics will still be displayed (maybe we could skip that for users that had opted in?).

Bug: None
Change-Id: If1cc12b2fc06f0d6237714c4f182367b1afdf9fb