Update patch set 4
Patch Set 4:

I tried this with a few random Python files in V8 and it changes some of them. I'm not against getting to well-formatted files, it'd actually be desired, but I'd not switch this on just globally.

As a minimum, this should be opt-in per project somehow. Maybe with some config in the directory tree?

Best would be to first format all python files in that project before turning on the option. Otherwise the formatting first happens on the first unrelated change to a file by some random author. Many of those authors will wonder what this is about...

Patch-set: 4
Reviewer: Gerrit User 1115956 <1115956@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>