Update patch set 1
Patch Set 1:

Fast work. LGTM though I'm not sure we should merge now.

I was thinking, it _might_ make sense to not stop the logging till after we fix the bug, if we can fix it this week. The reasoning being that this provides a great way to validate the bug fix. If anyone does still encounter it after we think we've fixed it, we'll get the headers & report rather than a potential new bug, etc. But I hate to have people doing this for weeks more if we don't need it.

Thoughts?

Patch-set: 1
Reviewer: Gerrit User 1303432 <1303432@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>