Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Fast work. LGTM though I'm not sure we should merge now.
> > 
> > I was thinking, it _might_ make sense to not stop the logging till after we fix the bug, if we can fix it this week. The reasoning being that this provides a great way to validate the bug fix. If anyone does still encounter it after we think we've fixed it, we'll get the headers & report rather than a potential new bug, etc. But I hate to have people doing this for weeks more if we don't need it.
> > 
> > Thoughts?
> 
> That's a good point. Do you think we can fix it this week?
> Otherwise, we could stop logging and just print a message telling them that it is a known bug and we're working on it.

Andrii is currently on the hook for the remaining changes though we can see if you can help if necessary to accelerate? I believe we're unblocked from an API perspective so if you guys have time this week we should be able to write what we think is the full fix.

Patch-set: 1
Reviewer: Gerrit User 1303432 <1303432@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>