Update patch set 8
Patch Set 8: Code-Review+1

I'm going to change my tune a bit and say LGTM.

I really do think we can and should do better with the README. But I really want this metrics change in so we can get a much clearer view of the errors and timings of git CL upload. While I'm sure landing this as is means we'll need at least one more README update later as I'm sure we'll tweak what we collect again in the future, it might be worth it.

If I've convinced you to rapidly iterate on the README now and still get this in today. That's okay with me too.

Patch-set: 8
Reviewer: Gerrit User 1303432 <1303432@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1