autoninja: use ninja in the autoninja.py directory
Since the original code did not specify ninja in absolute path,
it uses ninja in PATH.  If ninja does not exist in PATH, it fails
to execute ninja.

Having depot_tools in PATH is recommended and human users usually do
not see this issue.  However, it is not true for bots.  They usually
do not have depot_tools in PATH, and autoninja does not work.

Also, I believe using ninja in PATH is confusing.  When people execute
autoninja, I guess people would expect ninja in the same depot_tools
would be used.  However, if we get ninja from PATH, it is not
guaranteed to be so.

Bug: b/77176746
Change-Id: Iad8bd952dc1e34a9d303fd5b493c555156369a17