Update patch set 3
Patch Set 3:

Hm, ok... I think that might be confusing because e.g. recipes don't have the ninja binaries in their copy of depot_tools.

I think the right thing is probably to pick the one in depot_tools, but only if it exists. Otherwise try for the one in PATH.

Patch-set: 3