Update patch set 1
Patch Set 1:

I'll let dpranke review this one.

It's now VERY unclear to me why this functionality isn't just built into ninja, if all devs and bots are going to be using it all the time. Adding wrappers is still the thing that we do for tools that we have no control over, but AFAIK this is not the case for ninja.

To recap, IIUC:
  * We have added a python wrapper around ninja which:
    * parses ninja's command line arguments. This wrapper is not sourced from the same repo that ninja is, and so if the command line arguments change, this wrapper will need to be manually updated.
    * inspects the environment for GOMA-related envvars and then parses args.gn with regexes.
    * pulls `ninja` from the one checked into depot_tools. This makes the recipe bundle about 1MB+ larger than it needs to be, and is different than how we distribute other binary tools on the bots (which is not surprising, I think we have 5 ways to do this now. Not even gn is distributed the same way as ninja.)
    * modifies the command line argument of ninja to calculate it's -j value.
  * We have added a bat & sh wrapper around this python wrapper, adding an extra level of indirection.
    * This wrapper runs an additional python subprocess to generate a UUID for unclear purposes.
    * It also runs a followup script post_build_ninja_summary.py.

This reminds me distinctly of all the rietveld chrome extensions that people developed towards the end of its life; There were many features that people wanted from rietveld, but for various reasons decided the best thing to do was to monkeypatch the served html with javascript and CSS, instead of just changing the service to have the features that everyone wanted.

AFAIK, one of the main points of ninja is to be fast. Adding 4 extra processes to every invocation of ninja, 3 of which start python interpreters, seems to me like the opposite of fast.

Or maybe this autoninja thing actually makes sense and I'm the crazy one :). In any event, I'll let dirk make the call.

Patch-set: 1
Reviewer: Gerrit User 1001907 <1001907@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>