Update patch set 1
Patch Set 1: Code-Review+1

The CL lgtm.

Responses to other parts of iannucci's comments inline ...

> Patch Set 1:
> 
> I'll let dpranke review this one.
> 
> It's now VERY unclear to me why this functionality isn't just built into ninja, if all devs and bots are going to be using it all the time. Adding wrappers is still the thing that we do for tools that we have no control over, but AFAIK this is not the case for ninja.

Ninja is something we only sort of have control over, in the sense that most of the core committers to Ninja are related to Chrome, but it is treated as (and run as) an independent project. We only add things to Ninja that make sense when ninja is viewed as a standalone thing (like make or cmake).

The stuff we're working on here is very chromium+goma-specific, and doesn't make sense to be added to core Ninja.

> 
> To recap, IIUC:
>   * We have added a python wrapper around ninja which:
>     * parses ninja's command line arguments. This wrapper is not sourced from the same repo that ninja is, and so if the command line arguments change, this wrapper will need to be manually updated.

Ninja's command line has been stable for 6+ years and isn't likely to change any time soon.

>     * inspects the environment for GOMA-related envvars and then parses args.gn with regexes.
>     * pulls `ninja` from the one checked into depot_tools. This makes the recipe bundle about 1MB+ larger than it needs to be, and is different than how we distribute other binary tools on the bots (which is not surprising, I think we have 5 ways to do this now. Not even gn is distributed the same way as ninja.)

We should move ninja to be distributed via CIPD, and when we've done that we should switch depot_tools to pull in the binary via CIPD. Either way, it is correct for us to be using the binary distributed as part of depot_tools, rather than assuming it is installed elsewhere on the machine.

>     * modifies the command line argument of ninja to calculate it's -j value.
>   * We have added a bat & sh wrapper around this python wrapper, adding an extra level of indirection.

The bat wrapper, at least, is unavoidable for now, I think. We could probably get rid of the sh wrapper, but that probably has relatively low overhead as well.

>     * This wrapper runs an additional python subprocess to generate a UUID for unclear purposes.

We should have comments making clear what the UUID is for. I don't know why that would need to be done in a separate subprocess.

>     * It also runs a followup script post_build_ninja_summary.py.
> 
> This reminds me distinctly of all the rietveld chrome extensions that people developed towards the end of its life; There were many features that people wanted from rietveld, but for various reasons decided the best thing to do was to monkeypatch the served html with javascript and CSS, instead of just changing the service to have the features that everyone wanted.
> 
> AFAIK, one of the main points of ninja is to be fast. Adding 4 extra processes to every invocation of ninja, 3 of which start python interpreters, seems to me like the opposite of fast.

I agree with this observation, we should be making sure the invocation is fast. Also it's not clear to me why we'd need more than 1 (or maybe 2) invocations of python.

> 
> Or maybe this autoninja thing actually makes sense and I'm the crazy one :). In any event, I'll let dirk make the call.

The "long term" vision for this that we've been bandying about is to restructure ninja so that it can be distributed as a library + a command-line frontend to the library. At that point, we'd then consider having a binary that we ran that *wasn't* ninja that integrated directly with goma (and possibly GN) which would get rid of most of these wrappers and indirections and also provide other benefits.

Patch-set: 1
Reviewer: Gerrit User 1003232 <1003232@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1