[cipd] Bump CIPD client and vpython versions.
The most notable change is in how the integrity of the tag cache file is
checked. Once deployed, existing tag cache files will be considered invalid and
wiped.

The only observable side effect is the following line in the log:
"can't deserialize tag cache - no sha256 is recorded in the file"

R=tandrii@chromium.org, iannucci@chromium.org

Change-Id: I7ea300a4e7ad8be0c1d42ae561c5202420d2db62