Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> If nothing else this looks weird. It means that the command that is executed (which is printed out before running it) is:
> 
> "c:\src\depot_tools\ninja.exe" "-C" "out\release" "base" "-j" "960"
> 
> I mean, that's harmless, but I'm not sure it's improving anything.

Is following a valid command?
> ninja -C "out\A B" "C D"
Then, I guess it might not work on autoninja without this patch.

As far as I understand, ninja do not want to have a feature specific to certain
platform or build targets.  However, we want to evolve it to have features useful for
build performance improvement.  We need to have a wrapper for ninja, then.
Autoninja is currently in the such position, and I want people to use autoninja instead
of ninja 

From that stand point, I expect autoninja should behave as a replacement of ninja for any
command line arguments.  Otherwise, it blocks people to migrate to autoninja, and I am afraid
that such bad reputation ("XXX worked ninja but does not work for autoninja.  That is why
I do not use autoninja.") stays longer than we think.


Also, this CL includes the fix for 
https://bugs.chromium.org/p/chromium/issues/detail?id=902930#c7
In previous code, it did not have "eval".  Then, the command with quotes seems to
treated as-is, and bash cannot find the command with quotes.  With eval, bash evaluates
the command line, and can recognize quotes is not a part of a file name.

Patch-set: 1
Reviewer: Gerrit User 1107806 <1107806@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>