autoninja: quote the arguments when needed.
On Windows, without quote, whole return value would be treated as
a command, and the execution would fail.
On Linux and Mac, without quote, if depot_tools is settled under
directories with ' ', command execution would fail because paths are
separated in a wrong way.

To make such a return value work on Linux and Mac, the shell script
started to use eval.

Additional change:
- leave relative path as-is.

Bug: 902930
Change-Id: I9bb74585294af565988c0b844b6b113a5c685530