Update patch set 1
Patch Set 1:

I own this file (I've been the chief maintainer for a few years) so I'm definitely the right person to review it and approve it. I'll take a closer look at it later but I want to address one misconception that I see in the description and first comments: this package is used for building Chrome for Windows on Windows and on Linux. At least, it is used by Google employees, and our build machines. So, the ", even from Linux." and "This is required for FFmpeg rolls." parts of the description are erroneous. This change is needed so that our build machines can build ARM64 Win32, full stop.

After this is landed I will need to create a new package, probably with a new SDK, new version of VS header files, etc., but it's about time to do that again anyway.

Patch-set: 1
Reviewer: Gerrit User 1128439 <1128439@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>