Update patch set 2
Patch Set 2:

> I didn't see anything in my ../depot_tools/win_toolchain and third_party/depot_tools/win_toolchain folders, and from other errors it was clear it was using the installed SDK and not a packaged blob during my local builds. That, combined with the ffmpeg roll docs' insistence that the win configs must be updated from Linux, led to these comments. Now that I understand a bit better, I'll get these odd comments scrubbed out. :)

Yet, it's confusing. It's controlled by the DEPOT_TOOLS_WIN_TOOLCHAIN environment variable which is set to 1 for Google employees (and our build machines) and to 0 for everybody else. We'd love to set it to 1 for everybody but VS licensing requirements make that impossible.

Patch-set: 2
Reviewer: Gerrit User 1128439 <1128439@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>