Update patch set 2
Patch Set 2:

As the comment says, the quoting of all of the arguments was put in to fix a problem with handling of the ^ character. In particular, this usage of ninja is quite handy:

autoninja -C out\default ..\..\base\win\enum_variant.cc^^

This tells ninja to build the outputs of enum_variant.cc. Without the quoting of the arguments this requires something like 16 or 32 ^ characters to deal with all of the parsers stripping off escaped characters.

I'm still not sure what the benefit of quoting the arguments is - maybe we should leave things alone until we need the quoting?

Patch-set: 2
Reviewer: Gerrit User 1128439 <1128439@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>