[DevTools] Add a fetch config for the inspector_protocol.
This will facilitate a more standalone development process.
We don't have a DEPS file there yet, but I'm planning
to add a simple one, which for starters will
have gunit. I'll model this mostly after breakpad.
I've tested this fetch config and it appears to work great:

$ mkdir testfetch
$ cd testfetch
$ fetch inspector_protocol

After this there's an src directory in testfetch which
has our sources.

Change-Id: I482d65b3d8e50e9e5a98c4affc7b487a5bd6df53