Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 1: Commit-Queue+1
> > > 
> > > Testing if CQ+1 goes away when I edit description.
> > 
> > It does.
> 
> WAI or not?

Both... For this repo based on configuration it is WAI. However I later tested this in chromium/src with Andrii where it should persist and it also didn't there. So there's two takeaways:

(1) We should improve the config of this repo so that CQ+1 persists with a non-code change like editing description.

(2) The config that should ensure this isn't working and that should probably be a bug in Gerrit. Need to look into it more. This is what should work: https://chromium.googlesource.com/chromium/src/+blame/refs/meta/config/project.config#14

Patch-set: 2
Reviewer: Gerrit User 1303432 <1303432@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>