Update patch set 1
Patch Set 1:

(1 comment)

Hi Robbie, unclear who send this to, you seem to have added such a file recently. Maybe you can help or suggest a better reviewer?

We're using gsutil.py programmatically in some of our new tools. Unfortunately on Windows, we can't just do a PATH search for gsutil.py because it invokes the OS .py handler, which isn't set up to run the script. Most of the other programs in depot_tools has a .bat variant that is super easy to use, so if we add one for gsutil.py, that would be straightforward to use. The alternative is to use the python in the PATH and then reimplement a PATH search for depot_tools so we can pass the absolute path to gsutil.py to python, but that's more code. I think this is the simpler solution. It's slightly inconvenient that the program is called gsutil.py instead of gsutil, but that does avoid it conflicting with the google cloud sdk's gsutil program. I don't particularly want to require our developers to also install the google cloud sdk when depot_tools should do just fine and they already have it.

Patch-set: 1
Reviewer: Gerrit User 1001907 <1001907@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1116069 <1116069@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1238297 <1238297@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1183500 <1183500@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false